requirements.txt
# ── Core AI / CV ──────────────────────────────────────────────────────────────
ultralytics>=8.0.0          # YOLOv8 detection
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
timm>=0.9.0                 # EfficientNet / ViT classifiers

# ── Data Science ──────────────────────────────────────────────────────────────
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.66.0
scipy>=1.11.0

# ── Report Generation ─────────────────────────────────────────────────────────
reportlab>=4.0.0            # PDF report generation
fpdf2>=2.7.0                # Alternative PDF lib
openpyxl>=3.1.0             # Excel cost reports

# ── API / Web ─────────────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6     # File upload support
aiofiles>=23.0.0

# ── Utils ─────────────────────────────────────────────────────────────────────
PyYAML>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
filterpy>=1.4.5
supervision>=0.18.0
albumentations>=1.3.0       # Advanced augmentation
